Server crash fix
- Seperated client-only config screen registration from common code so
it isn't loaded on the server, fixes #3
- Slightly tweaked item nbt condition so comparing itemstack counts
works properly
	- Updated level_30_enchant example action due to this

+++ src/main/java/highfox/inventoryactions/action/condition/ItemNbtCondition.java
@@ -26,7 +26,7 @@ public class ItemNbtCondition extends ItemSourcingCondition {
 	public boolean test(IActionContext context) {
 		ItemStack stack = this.source.get(context);
 		CompoundTag itemNbt = stack.hasTag() ? stack.getTag() : new CompoundTag();
-		itemNbt.putByte("Count", (byte)stack.getCount());
+		itemNbt.putInt("Count", stack.getCount());
 
 		return NbtUtils.compareNbt(this.nbt, itemNbt);
 	}
